refactor(desktop): split chatinput into hooks/parts/lib
